# openapi.yaml
openapi: 3.0.3
info:
  title: Lift Log API
  description: |
    Lift Log は筋力トレーニングの記録・管理を行うWebアプリケーションです。
    このAPIは JWT認証を使用し、すべてのAPIリクエストにはAuthorizationヘッダーが必要です。
  version: 1.0.0
  contact:
    name: Lift Log Support
    email: support@liftlog.example.com
servers:
  - url: http://localhost:8080/api
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
  schemas:
...
    WorkoutRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ワークアウト記録ID
        workoutDayId:
          type: string
          format: uuid
          description: ワークアウト日ID
        exerciseId:
          type: string
          format: uuid
          description: エクササイズID
        notes:
          type: string
          nullable: true
          description: メモ
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkoutRecordRequest:
      type: object
      required:
        - workoutDayId
        - exerciseId
      properties:
        workoutDayId:
          type: string
          format: uuid
          description: ワークアウト日ID
        exerciseId:
          type: string
          format: uuid
          description: エクササイズID
        notes:
          type: string
          description: メモ
...
paths:
...
  /workout-records:
    get:
      tags:
        - Workout Records
      summary: ワークアウト記録一覧取得
      description: 認証されたユーザーのすべてのワークアウト記録を取得します
      responses:
        '200':
          description: ワークアウト記録一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutRecord'
        '401':
          description: 認証失敗
    post:
      tags:
        - Workout Records
      summary: ワークアウト記録作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutRecordRequest'
      responses:
        '201':
          description: ワークアウト記録作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutRecord'
  /workout-records/{id}:
    get:
      tags:
        - Workout Records
      summary: ワークアウト記録詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ワークアウト記録詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutRecord'
        '404':
          description: ワークアウト記録が見つからない
    put:
      tags:
        - Workout Records
      summary: ワークアウト記録更新
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutRecordRequest'
      responses:
        '200':
          description: ワークアウト記録更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutRecord'
        '404':
          description: ワークアウト記録が見つからない
    delete:
      tags:
        - Workout Records
      summary: ワークアウト記録削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: ワークアウト記録削除成功
        '404':
          description: ワークアウト記録が見つからない
  /workout-days/{workoutDayId}/workout-records:
    get:
      tags:
        - Workout Records
      summary: ワークアウト日別記録一覧取得
      parameters:
        - name: workoutDayId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ワークアウト記録一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutRecord'
    post:
      tags:
        - Workout Records
      summary: ワークアウト記録作成（パス指定）
      parameters:
        - name: workoutDayId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutRecordRequest'
      responses:
        '201':
          description: ワークアウト記録作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutRecord'
...